#13.2 - MainActivity and ProductList finished (including proper loading/pbar)
