+++ alodjinha/app/src/main/java/com/abmm/b2w/alodjinha/http_module/ILodjinhaApi.java
@@ -6,6 +6,7 @@ import com.abmm.b2w.alodjinha.model.Product;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
+import retrofit2.http.POST;
 import retrofit2.http.Path;
 import retrofit2.http.Query;
 
@@ -25,4 +26,6 @@ public interface ILodjinhaApi {
 
     @GET("/produto/{id}")
     Call<Product> getProductById(@Path("id") int id);
+
+
 }
