+++ alodjinha/app/src/main/res/drawable/progress_bar_orange.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/progress_bar_orange_outerring"/>
+</layer-list>
