+++ alodjinha/app/src/main/res/drawable/progress_bar_orange_outerring.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<rotate xmlns:android="http://schemas.android.com/apk/res/android"
+    android:pivotX="50%" android:pivotY="50%" android:fromDegrees="0"
+    android:toDegrees="720">
+
+    <shape android:shape="ring" android:innerRadiusRatio="3"
+        android:thicknessRatio="8" android:useLevel="false">
+
+        <size android:width="48dip" android:height="48dip" />
+
+        <gradient android:type="sweep" android:useLevel="false"
+            android:startColor="@android:color/transparent" android:centerColor="@color/tomato"
+            android:centerY="0.50" android:endColor="@color/tomato" />
+
+    </shape>
+</rotate>
